added new routers (chat and course), refactoring app.py, stll in progress, pending Google Style comments once completed
